Keep only the patient list with med. history (#3535)
* Keep only the patient list with med history

- Remove `Patient list`
- Rename `Patient list with med. history` to `Patient line list`

* Update execpected_usages of `let!`
